requirements.txt
ttkbootstrap
openpyxl
pandas
fpdf
reportlab
typeguard

requests
tkcalendar
Pillow
colorama
tabulate
bcrypt
pycryptodome
python-dateutil
